Add getter for GVDB table contents
The xdg-desktop-portal project is using GVDB as the basis for the
permissions store.

Each permission table is stores as a block of bytes inside another GVDB
table, in order to isolate the permissions of each application.

Currently, xdg-desktop-portal uses GVDB as a copy-paste library; the
only missing API to avoid that, and switch GVDB to a proper subproject,
is a way to extract the contents of a GVDB table into a GBytes buffer.