build: Add -Dtests configure option
While the test suite is good, it introduces a dependency loop when
building glib.git: glib.git pulls in gvdb.git as a subproject so that it
can be statically built into `libgio-2.0.so`. The new unit tests in
`gvdb/tests/meson.build` call `dependency('gio-2.0')`. GLib overrides
this dependency, but it can’t do so until it’s defined how
`libgio-2.0.so` is built, but that depends on GVDB, so GVDB’s
`dependency('gio-2.0')` is called first. This means the dependency can’t
be found.

It’s not strictly a dependency loop (GVDB tests depend on libgio and the
GVDB sources; libgio depends on the GVDB sources; nothing depends on the
GVDB tests) but acts as one in effect because the `meson.build` for the
tests can’t be split from the `meson.build` for the GVDB sources.

So, add `-Dtests={true,false}` so GLib can disable the tests when using
GVDB as a submodule.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>